1、新增：工程提交 2、优化：新建工程，可以新建在已有文件夹中

+++ SuperMap.ZS/SuperMap.ZS.Startup/DataManager/NewWorkspaceCtrlAction.cs
@@ -7,6 +7,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
 using System.Text;
 using System.Windows.Forms;
 
@@ -16,6 +17,7 @@ namespace SuperMap.ZS.Startup
     {
         private int m_TotleCount = 0;
         private int m_Count = 0;
+        private string m_SelectFolderPath;
 
 #pragma warning disable CS0612 // 类型或成员已过时
         public NewWorkspaceCtrlAction()
@@ -45,22 +47,40 @@ namespace SuperMap.ZS.Startup
                 };
                 if(folder.ShowDialog() == DialogResult.OK)
                 {
+                    m_SelectFolderPath = folder.SelectedPath;
                     string[] files = Directory.GetFiles(folder.SelectedPath);
                     if (files.Length > 0)
                     {
                         UIMessageBox box = new UIMessageBox(MessageBoxUIStyle.iDesktopStyle);
                         if (box.Show("该文件夹不为空，是否覆盖？", "提示", MessageBoxButtons.OKCancel, MessageBoxIcon.Question, MessageBoxDefaultButton.Button1) == DialogResult.OK)
                         {
-                            foreach (string file in files)
+                            if (!string.IsNullOrEmpty(Desktop.Application.ActiveApplication.Workspace.ConnectionInfo.Server))
                             {
-                                File.Delete(file);
+                                Desktop.Application.ActiveApplication.Workspace.Closed += Workspace_Closed;
+                                Desktop.Application.ActiveApplication.Workspace.Close();
+                            }
+                            else
+                            {
+                                foreach (string file in files)
+                                {
+                                    File.Delete(file);
+                                }
+                                m_Count = 0;
+                                FTPController ftp = new FTPController();
+                                ftp.OnUpdateComplete += Ftp_OnUpdateComplete;
+                                ftp.UpdateOriginalData(folder.SelectedPath);
+                                m_TotleCount = ftp.UpdateFilesCount;
+                                CommonPars.CurrentWorkspaceDir = folder.SelectedPath;
                             }
                         }
                         else
                         {
                             return;
                         }
-
+                    }
+                    else
+                    {
+                        m_Count = 0;
                         FTPController ftp = new FTPController();
                         ftp.OnUpdateComplete += Ftp_OnUpdateComplete;
                         ftp.UpdateOriginalData(folder.SelectedPath);
@@ -76,6 +96,29 @@ namespace SuperMap.ZS.Startup
             }
         }
 
+        private void Workspace_Closed(object sender, WorkspaceClosedEventArgs args)
+        {
+            try
+            {
+                string[] files = Directory.GetFiles(m_SelectFolderPath);
+                foreach (string file in files)
+                {
+                    File.Delete(file);
+                }
+                m_Count = 0;
+                FTPController ftp = new FTPController();
+                ftp.OnUpdateComplete += Ftp_OnUpdateComplete;
+                ftp.UpdateOriginalData(m_SelectFolderPath);
+                m_TotleCount = ftp.UpdateFilesCount;
+                CommonPars.CurrentWorkspaceDir = m_SelectFolderPath;
+                Desktop.Application.ActiveApplication.Workspace.Closed -= Workspace_Closed;
+            }
+            catch (Exception ex)
+            {
+                Log.OutputBox(ex);
+            }
+        }
+
         private void Ftp_OnUpdateComplete(object sender, System.ComponentModel.AsyncCompletedEventArgs e)
         {
             try
@@ -88,6 +131,12 @@ namespace SuperMap.ZS.Startup
                         FileInfo info = new FileInfo(file);
                         if (info.Extension.Contains("smwu") || info.Extension.Contains("sxwu"))
                         {
+                            //改名
+                            string name = CommonPars.CurrentWorkspaceDir.Substring(CommonPars.CurrentWorkspaceDir.LastIndexOf("\\") + 1);
+                            string path = CommonPars.CurrentWorkspaceDir + "\\" + name + info.Extension;
+                            info.MoveTo(path);
+
+                            //打开
                             WorkspaceConnectionInfo wsinfo = new WorkspaceConnectionInfo()
                             {
                                 Server = info.FullName,
@@ -95,6 +144,8 @@ namespace SuperMap.ZS.Startup
                             };
                             if (Desktop.Application.ActiveApplication.Workspace.Open(wsinfo))
                             {
+                                Desktop.Application.ActiveApplication.Workspace.Caption = name;
+                                Desktop.Application.ActiveApplication.Workspace.Save();
                                 Desktop.Application.ActiveApplication.CreateSceneWindow(Desktop.Application.ActiveApplication.Workspace.Scenes[0]);
                                 Desktop.Application.ActiveApplication.Output.Output("工作空间【" + info.Name.Split('.')[0] + "】已创建！", InfoType.Information);
                             }
